reduce surface area of get_enabled_materialized_columns
