+++ posthog/clickhouse/materialized_columns.py
@@ -1,6 +1,4 @@
-from datetime import timedelta
-
-from posthog.cache_utils import cache_for
+from posthog.models.instance_setting import get_instance_setting
 from posthog.models.property import PropertyName, TableColumn, TableWithProperties
 from posthog.settings import EE_AVAILABLE
 
@@ -8,19 +6,20 @@ from posthog.settings import EE_AVAILABLE
 ColumnName = str
 TablesWithMaterializedColumns = TableWithProperties
 
+
 if EE_AVAILABLE:
-    from ee.clickhouse.materialized_columns.columns import get_materialized_columns
+    from ee.clickhouse.materialized_columns.columns import get_enabled_materialized_columns
+
+    def get_materialized_column_for_property(
+        table: TablesWithMaterializedColumns, table_column: TableColumn, property_name: PropertyName
+    ) -> str | None:
+        if not get_instance_setting("MATERIALIZED_COLUMNS_ENABLED"):
+            return None
+
+        return get_enabled_materialized_columns(table).get((property_name, table_column))
 else:
 
-    def get_materialized_columns(
-        table: TablesWithMaterializedColumns,
-        exclude_disabled_columns: bool = False,
-    ) -> dict[tuple[PropertyName, TableColumn], ColumnName]:
-        return {}
-
-
-@cache_for(timedelta(minutes=15))
-def get_enabled_materialized_columns(
-    table: TablesWithMaterializedColumns,
-) -> dict[tuple[PropertyName, TableColumn], ColumnName]:
-    return get_materialized_columns(table, exclude_disabled_columns=True)
+    def get_materialized_column_for_property(
+        table: TablesWithMaterializedColumns, table_column: TableColumn, property_name: PropertyName
+    ) -> str | None:
+        return None
