+++ posthog/test/base.py
@@ -30,7 +30,6 @@ from rest_framework.test import APITestCase as DRFTestCase
 from posthog import rate_limit, redis
 from posthog.clickhouse.client import sync_execute
 from posthog.clickhouse.client.connection import ch_pool
-from posthog.clickhouse.materialized_columns import get_materialized_columns
 from posthog.clickhouse.plugin_log_entries import TRUNCATE_PLUGIN_LOG_ENTRIES_TABLE_SQL
 from posthog.cloud_utils import TEST_clear_instance_license_cache
 from posthog.models import Dashboard, DashboardTile, Insight, Organization, Team, User
@@ -566,22 +565,14 @@ def stripResponse(response, remove=("action", "label", "persons_urls", "filter")
     return response
 
 
-def default_materialised_columns():
+def cleanup_materialized_columns():
     try:
+        from ee.clickhouse.materialized_columns.columns import get_materialized_columns
         from ee.clickhouse.materialized_columns.test.test_columns import EVENTS_TABLE_DEFAULT_MATERIALIZED_COLUMNS
     except:
         # EE not available? Skip
-        return []
+        return
 
-    default_columns = []
-    for prop in EVENTS_TABLE_DEFAULT_MATERIALIZED_COLUMNS:
-        column_name = get_materialized_columns("events")[(prop, "properties")]
-        default_columns.append(column_name)
-
-    return default_columns
-
-
-def cleanup_materialized_columns():
     def optionally_drop(table, filter=None):
         drops = ",".join(
             [
@@ -593,7 +584,10 @@ def cleanup_materialized_columns():
         if drops:
             sync_execute(f"ALTER TABLE {table} {drops} SETTINGS mutations_sync = 2")
 
-    default_columns = default_materialised_columns()
+    default_columns = [
+        get_materialized_columns("events")[(prop, "properties")] for prop in EVENTS_TABLE_DEFAULT_MATERIALIZED_COLUMNS
+    ]
+
     optionally_drop("events", lambda name: name not in default_columns)
     optionally_drop("person")
     optionally_drop("groups")
